Deterministically build a dependency deb package
This way, Docker can detect there is no change in the package and not
reinstall dependencies. This uses strip-nondeterminism which is not yet
widely available. Maybe this could be done with more common
tools. Instead of using equivs, it uses directly dpkg-deb. This has the
drawback of not handling correctly arch-specific
build-dependencies. This is fixed in the next commit.